backend/requirements.txt
django-cors-headers
Django
djangorestframework
djangorestframework-simplejwt
gTTS
transformers
torch
sentencepiece
pdfplumber
python-docx